eslint --fix and pass API tests
